requirements.txt
aiogram==3.3.0
py7zr==0.21.2
aiohttp==3.9.4
fastapi==0.111.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
internetarchive==1.10.1